v1.0.1: add option to redirect video URL through non-HTTPS endpoint
"force_http_video_url" is enabled by default,
because I'm choosing to prioritize support
for using ExoAirPlayer over that of Chromecast.

older Android devices require this feature to be enabled;
it (somehow) serves as a workaround for a failed TLS handshake.

Chromecast requires this feature to be disabled;
the user will need to change the value of this option,
in order to send videos to Chromecast.